PHPCS 3.x compat: add helper class & implement use
A number of PHPCS 2.x classes have been split out over several classes in PHPCS 3.x.
Aliasing the new classname to the old will therefore not work.

This helper class acts as a compatibility layer for these cases.
The helper class only addresses the functionality needed by PHPCompatibility and is not exhaustive.